# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AlustarsCRM API",
    "version": "1.0.0",
    "description": "AlustarsCRM 客户关系管理系统 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/settings/master-data/export": {
      "get": {
        "operationId": "get-settings-master-data-export",
        "summary": "Master Data Export API Route",
        "description": "",
        "tags": [
          "Settings"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customers": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "customer_contacts": {
                      "type": "object"
                    },
                    "suppliers": {
                      "type": "object"
                    },
                    "supplier_contacts": {
                      "type": "object"
                    },
                    "supplier_bank_accounts": {
                      "type": "object"
                    },
                    "products": {
                      "type": "object"
                    },
                    "app_config": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "customers",
                    "customer_contacts",
                    "suppliers",
                    "supplier_contacts",
                    "supplier_bank_accounts",
                    "products",
                    "app_config"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/shipments/{id}/items": {
      "get": {
        "operationId": "get-shipments-{id}-items",
        "summary": "",
        "description": "",
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "shipment": {
                            "type": "object"
                          },
                          "order_item": {
                            "type": "object"
                          },
                          "quantity": {
                            "type": "number"
                          },
                          "packages": {
                            "oneOf": [
                              {
                                "type": "object"
                              },
                              {
                                "type": "object"
                              }
                            ]
                          },
                          "gross_weight": {
                            "type": "object"
                          },
                          "net_weight": {
                            "type": "object"
                          },
                          "volume": {
                            "type": "object"
                          },
                          "package_length": {
                            "type": "object"
                          },
                          "package_width": {
                            "type": "object"
                          },
                          "package_height": {
                            "type": "object"
                          },
                          "part_number": {
                            "oneOf": [
                              {
                                "type": "object"
                              },
                              {
                                "type": "object"
                              }
                            ]
                          },
                          "product_code": {
                            "type": "object"
                          },
                          "product_name": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "id",
                          "shipment",
                          "order_item",
                          "quantity",
                          "packages",
                          "gross_weight",
                          "net_weight",
                          "volume",
                          "package_length",
                          "package_width",
                          "package_height",
                          "part_number",
                          "product_code",
                          "product_name"
                        ]
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-shipments-{id}-items",
        "summary": "",
        "description": "",
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "collectionId": {
                            "type": "string"
                          },
                          "collectionName": {
                            "type": "object"
                          },
                          "expand": {
                            "oneOf": [
                              {
                                "type": "object"
                              },
                              {
                                "type": "object",
                                "additionalProperties": {
                                  "type": "object"
                                }
                              }
                            ]
                          },
                          "id": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "collectionId",
                          "collectionName",
                          "id"
                        ]
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}